GCounter and Counter CRDTs working, some bugs (count infrequently incorrect)

+++ pserver-crdt/src/main/java/de/tuberlin/pserver/counters/CounterCRDT.java
@@ -1,8 +1,6 @@
 package de.tuberlin.pserver.counters;
 
-import de.tuberlin.pserver.runtime.DataManager;
-
 public interface CounterCRDT {
 
-    int getCount();
+    long getCount();
 }
